+++ src/sub-clients/contentPartners/model/ContentPartner.ts
@@ -1,5 +1,4 @@
 import moment from 'moment';
-import { AgeRange } from '../../common/model/AgeRange';
 import { DistributionMethod } from '../../common/model/DistributionMethod';
 import { Link } from '../../common/model/LinkEntity';
 import { LegalRestriction } from '../../legalRestrictions/model/LegalRestriction';
@@ -13,7 +12,6 @@ export interface ContentPartner {
   id: string;
   name: string;
   official: boolean;
-  ageRange?: AgeRange;
   currency?: string;
   legalRestriction?: LegalRestriction;
   distributionMethods: DistributionMethod[];