+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Removed unnecessary `console.log`s
+
 ## [8.6.5] - 2020-04-08
 
 ### Fixed